SPL: refactor FDT FIT loading
Split the spl_fit_select_fdt() function that determines the right
configuration _and_ finds the matching DTB into:
1) a function that just returns the node for the matching configuration
2) a function that returns the image data for a given property in that
   configuration node, additionally using a certain index into a list of
   arguments (spl_fit_select_index()). This function is a generalized
   version of the existing spl_fit_select_fdt() function.

This patch introduces no functional changes, it just refactors the code
to allow reusing it later.

(diff is overly clever here and produces a hard-to-read patch, so I
recommend to throw a look at the result instead).

Signed-off-by: Andre Przywara <andre.przywara@arm.com>