ARM: boot0 hook: remove macro, include whole header file
For prepending some header area to U-Boot images we were so far including
a header file, in which there was a macro definition.
This works fine for code which covers just a few lines.
However adding more code quickly gets messy, so let's just drop that
intermediate macro and let the #include actually insert the code
directly.
This converts the callers and the callees, but doesn't change anything
at this point.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>